refactor: change sort_order column type from unsignedSmallInteger to integer in multiple migration files

+++ database/migrations/2026_05_08_231204_create_event_assignments_table.php
@@ -13,7 +13,7 @@ return new class extends Migration
             $table->id();
             $table->foreignIdFor(Event::class)->constrained()->cascadeOnDelete();
             $table->foreignId('boss_id')->nullable()->constrained('bosses')->cascadeOnDelete();
-            $table->unsignedSmallInteger('sort_order')->nullable();
+            $table->integer('sort_order')->nullable();
             $table->string('left_type')->nullable();
             $table->text('left_value')->nullable();
             $table->string('right_type')->nullable();
