+++ database/migrations/2026_08_28_140801_rename_bosses_encounter_order_to_sort_order.php
@@ -16,7 +16,7 @@ return new class extends Migration
         });
 
         Schema::table('bosses', function (Blueprint $table) {
-            $table->unsignedSmallInteger('sort_order')->nullable(false)->change();
+            $table->integer('sort_order')->nullable(false)->change();
         });
     }
