+++ database/migrations/2026_08_28_151523_rename_guild_ranks_position_to_sort_order.php
@@ -20,7 +20,7 @@ return new class extends Migration
         });
 
         Schema::table('guild_ranks', function (Blueprint $table) {
-            $table->unsignedSmallInteger('sort_order')->nullable(false)->default(0)->change();
+            $table->integer('sort_order')->nullable(false)->default(0)->change();
         });
     }
